Upgrade cppcheck to 2.1 and apply patches (#2313)
* Upgrade cppcheck to 2.1 and apply patches

This change upgrades cppcheck to 2.1 and applies two patches from the unreleased 2.2 version to fix two false positives. The unreleased code currently runs much slower than 2.1, so we should wait until it is released to fully switch.

This commit also includes the following cppcheck related changes
* Removes some unneeded suppressions
* Changes ConfigurationNotChecked suppressions to per-file, which seems to work better
* Turns on the --force option so all configurations are checked
* Removes the kex variable from s2n_client_key_exchange.c as it was triggering  (style:shadowArgument) Local variable 'kex' shadows outer argument
* Added suppressions for false positives of sizeofDivisionMemfunc
* Changes expected_size in s2n_rsa.c to an int to allow for the GUARD macro to check it properly, as size_t cannot be negative
* Replaces a PRECONDITION_POSIX check in s2n_safety.c as the function return type is bool, so a -1 would overflow to true
* Changes a string literal to a pointer in s2n_safety_test.c to fix (warning:literalWithCharPtrCompare) String literal compared with variable 'NULL'. Did you intend to use strcmp() instead?
* Added a sizeofDereferencedVoidPointer suppression in a couple cases where macro was calling sizeof on a void pointer when CBMC was in use
* Adds the current directory to the include path to allow cppcheck to find the correct headers

* Fix linting issue with cppcheck suppression comment

* Adding --no-commit to cherry-pick to avoid "empty ident name not allowed"

* Updating safety sidetrail patch

* Fixing safety sidetrail patch

* Adding tests for null inputs to s2n_constant_time_equals

* Fixing the loop index for the s2n_constant_time_quals CBMC proof and renaming the name to match the function

* Replacing notnull_check with more specific check to remove need for sizeOfDereferencedVoidPointer suppression

+++ crypto/s2n_rsa.c
@@ -99,7 +99,7 @@ static int s2n_rsa_encrypt(const struct s2n_pkey *pub, struct s2n_blob *in, stru
 static int s2n_rsa_decrypt(const struct s2n_pkey *priv, struct s2n_blob *in, struct s2n_blob *out)
 {
     unsigned char intermediate[4096];
-    const size_t expected_size = s2n_rsa_encrypted_size(priv);
+    const int expected_size = s2n_rsa_encrypted_size(priv);
 
     GUARD(expected_size);
     S2N_ERROR_IF(expected_size > sizeof(intermediate), S2N_ERR_NOMEM);
